feat: integrate BullMQ for booking expiration management
- Added BullMQ dependency for job scheduling.
- Implemented BookingsExpireService to handle expiration of pending bookings.
- Created BookingsJobsScheduler to schedule the expiration job.
- Developed BookingsProcessor to process expiration jobs.
- Updated BookingsService to include booking creation and confirmation logic.
- Enhanced BookingsController with endpoints for creating, confirming, and canceling bookings.
- Added DTOs for booking operations: CreateBookingDto, ConfirmBookingDto, and CancelBookingDto.
- Modified ListingsService to simplify listing retrieval.
- Updated availability block service to include authorization checks.
- Introduced Docker Compose configuration for application and database services.
- Refactored various modules to improve code structure and maintainability.

+++ src/modules/availabilityBlocks/availabilityBlocks.service.ts
@@ -1,4 +1,4 @@
-import { BadRequestException, Injectable, NotFoundException } from '@nestjs/common';
+import { BadRequestException, Injectable, NotFoundException, UnauthorizedException } from '@nestjs/common';
 import { InjectRepository } from '@nestjs/typeorm';
 import { AvailabilityBlocks } from './availabilityBlocks.entity';
 import { Repository } from 'typeorm';
@@ -9,15 +9,17 @@ export class AvailabilityBlockService {
     constructor(@InjectRepository(AvailabilityBlocks) private availabilityBlockService: Repository<AvailabilityBlocks>, private listingService: ListingsService) {}
 
     async createAvailabilityBlock(hostId: number, listingId: number, startDate: Date, endDate: Date, reason: string) {
-        const listing = this.listingService.getListing(listingId, hostId)
+        const listing = await this.listingService.getListing(listingId)
         if (!listing) throw new NotFoundException('Listing not found')
+        if (listing.host.id != hostId) throw new UnauthorizedException('You are not the host of listing')
 
-        const AvailabilityBlock = await this.availabilityBlockService.create({
+        const AvailabilityBlock = this.availabilityBlockService.create({
             listing: { id: listingId },
             startDate,
             endDate,
             reason
         })
+
         return this.availabilityBlockService.save(AvailabilityBlock)
     }
 
@@ -39,7 +41,8 @@ export class AvailabilityBlockService {
         const availabilityBlock = await this.availabilityBlockService.findOne({ where: { id }, relations: { listing: true } })
         if (!availabilityBlock) throw new NotFoundException('Availability Block not found')
 
-        await this.listingService.getListing(availabilityBlock.listing.id, hostId)
+        const listing = await this.listingService.getListing(availabilityBlock.listing.id)
+        if (listing.host.id != hostId) throw new UnauthorizedException('You are not the host of listing')
 
         return await this.availabilityBlockService.remove(availabilityBlock)
     }
